Enable incremental document sync
We can reconstruct the file from the syntax tree, and then insert the
changes into that. This is slightly slower on the server side but can
give large improvements overall due to less json being sent to the
server.

If the document versions are sent out of order then a
textDocument/resync notification is sent to the client to request that
it sends the full document contents. This is not a standard notification
and requires client side handling.