Fix parsing of strings with escape characters.
When parsing strings escape characters were ignored which meant that if
a string contained a \" it would end the parsing of the string. We now
take into account \ when parsing a string and if the previous character
is a \ we don't end the string parsing if the next character is a ".