Documentation hover for AArch64
Parse the A64 ISA XML and generate documentation for each instruction,
this can then be used to show a documentation hover for the
instruction, the correct instruction is determined by looking at the
example assembly and parsing that and comparing the token types.

The expansion of the example assembly is slightly incorrect for some
instructions due to the use of { } in the instruction, currently the
documentation assumes all { } are for optional parts.