Handle non optional { } when parsing the documentation templates.
This still assumes that all optional { } are contained within an outer
most { }. If that's not the case the template will be expanded
incorrectly.

Remove the + character from register names as the parser will
currently split those up.

Sorted the documentation register list and remove duplicates this
should make the diffs in the future easier.