Add ability to run llvm-mca on code.
Add an extension method of asm/runAnalysis that will call llvm-mca on
the region (or file, if a region isn't provided). The architecture of
the file will be used. The CPU is configured with a map of
architecture to target cpu. If a cpu isn't specified for the
architecture then the host cpu will be used if the architecture is the
host architecture, otherwise generic will be used.