Document highlight based on the access type.
Rather than highlight references to something as Text for
everything attempt to determine the access type based on the
documentation we have. This will highlight registers as Read or Write
if this can be determined from the description of the register for the
instruction. If it can't be determined it will default back to a Text
reference.